requirements.txt
cairosvg==2.*
chess==1.*
flask==2.*
firebase-admin==6.*
pyyaml==6.*
